+++ apps/site/app/layout.tsx
@@ -1,11 +1,11 @@
 import { Analytics } from '@vercel/analytics/react'
 import type { Metadata } from 'next'
 import { GeistSans } from 'geist/font/sans'
-import { PackageInstallScript } from 'renoun/components'
+import { PackageInstallScript, Refresh } from 'renoun/components'
 
 export const metadata = {
-  title: 'renoun',
-  description: 'The Documentation Toolkit for React',
+  title: 'renoun - Elevate Your Design System Docs',
+  description: `The renoun toolkit uses your React framework to keep documentation polished, in sync, and on brand.`,
 } satisfies Metadata
 
 export default function RootLayout({
@@ -33,6 +33,7 @@ export default function RootLayout({
         />
         {children}
         <Analytics />
+        <Refresh />
       </body>
     </html>
   )