+++ apps/site/components/SiblingLink.tsx
@@ -169,12 +169,14 @@ async function resolveEntryMetadata(entry: FileSystemEntry) {
     return
   }
 
-  const metadataExport = await file.getExport('metadata').catch((error) => {
-    if (error instanceof FileExportNotFoundError) {
-      return undefined
-    }
-    throw error
-  })
+  const metadataExport = await file
+    .getNamedExport('metadata')
+    .catch((error) => {
+      if (error instanceof FileExportNotFoundError) {
+        return undefined
+      }
+      throw error
+    })
 
   if (metadataExport) {
     return metadataExport.getRuntimeValue()