+++ packages/mdx/src/remark/transform-relative-links.ts
@@ -2,7 +2,7 @@ import type { Root, Link } from 'mdast'
 import { visit } from 'unist-util-visit'
 
 /** Reformat all relative links that use ordered numbers and extensions. */
-export function transformRelativeLinks() {
+export default function transformRelativeLinks() {
   return (tree: Root) => {
     visit(tree, 'link', (node: Link) => {
       if (!/\d+.*\.(md|mdx)$/.test(node.url)) {