+++ packages/renoun/src/components/Theme/index.ts
@@ -1,3 +1,2 @@
 export { ThemeProvider } from './Theme.js'
 export { ThemeStyles } from './ThemeStyles.js'
-export { useThemePicker } from './use-theme-picker.js'