+++ packages/create-renoun/CHANGELOG.md
@@ -1,5 +1,17 @@
 # create-renoun
 
+## 1.2.1
+
+### Patch Changes
+
+- d0f3ab8: Fixes the working directory printed after configuring examples.
+
+## 1.2.0
+
+### Minor Changes
+
+- a862ea2: Updates all dependencies to latest version.
+
 ## 1.1.0
 
 ### Minor Changes